rename files for simplicity
