--- dynestyx/internal/__init__.py
@@ -1,1 +1,0 @@
-"""Internal helpers for simulator and inference implementations."""
